Refactor: Organize project structure
- Created `config`, `css`, and `sql` directories to better organize project files.
- Moved `db_config.php` to `config/`.
- Moved all CSS files (`.css`) to `css/`.
- Moved `barbershop_full_schema.sql` to `sql/`.
- Updated all file references in `index.php`, `login.php`, and `superadmin_dashboard.php` to reflect the new file paths.
