+++ registro/config/db_config.php
