+++ registro/sql/barbershop_full_schema.sql
