+++ registro/superadmin_dashboard.php
@@ -1,6 +1,6 @@
 <?php
 session_start();
-require_once 'db_config.php';
+require_once 'config/db_config.php';
 
 // 1. Proteger la página y verificar el rol de superadmin
 if (!isset($_SESSION['user_id']) || $_SESSION['user_rol'] !== 'superadmin') {
@@ -104,7 +104,7 @@ $admins_query = $mysqli->query("
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Super Admin Dashboard</title>
     <link href="https://fonts.googleapis.com/css2?family=Barlow:wght@400;600;700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="admin_styles.css">
+    <link rel="stylesheet" href="css/admin_styles.css">
 </head>
 <body>
     <div class="dashboard-container">
